feat(logger): add Logger interface and move slog logger setup in "slog.go" file
- Add Logger interface with base logging methods and zap fields
- Remove slog logger code to "slog.go" file
